pyproject.toml
[project]
name = "distiller"
version = "0.1.0"
description = "Claude Code Training Data Pipeline"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
s3 = [
    "boto3>=1.26.0",
]

[project.scripts]
distiller = "distiller.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"

[tool.hatch.build.targets.wheel]
packages = ["distiller", "scripts"]
